fix unsigned integer issues (WHY DIDN'T JAVA SUPPORT UNSIGNED NUMBERS!!!)

+++ src/javaforce/controls/s7/S7Data.java
@@ -7,7 +7,7 @@ package javaforce.controls.s7;
 
 public class S7Data {
   public byte block_type;
-  public byte block_number;
+  public short block_number;
   public byte data_type;
   public short offset;
   public short length;
